fix(slam): carry poses in double through surfel extraction and write-back (#351)
ProjectDB stores poses as double[16], but the old to_affine() returned
Affine3f, so every optimize/register run wrote back float-truncated poses
even when the solve did not move them.  Measured: optimize --no-plane-factors
changed 237/238 pose blobs by up to 1.43 µm — enough to shift voxel grid
and region-growing assignments, explaining the ~0.2 mm flatness_rms
determinism smell recorded in §9.8.

Fix: to_affine() now returns Affine3d (bit-exact round-trip);
to_array16(Affine3d) added alongside the kept Affine3f overload for
rendering callers.  FrameSurfels::world_pose widened to Affine3d.
All four write-back paths (PlaneGraphOptimizer, JointPairwiseRegistration,
optimize_sensor_poses, refine_sensor_poses) now compose and store Affine3d.
Per-point math in reconstruct.cpp and VTK rendering in render_view.cpp
still cast to float at the point of use, which is correct.

Tests:
- PosePrecision_RoundTrip_BitIdentical: to_array16(to_affine(arr)) == arr
  element-by-element (bit-exact, georeferenced-scale coordinates)
- PosePrecision_FloatOverload_WidensLossily: documents that the Affine3f
  overload is intentionally lossy (kept for rendering)
- ToAffineToArray16_RoundTrip_PreservesMatrix tightened from 1e-5 to 1e-15

Re-baseline after #351:
- HeavyDriftAlternatingRounds test: 3-round vs 1-round total error now
  differ by ~4 nm (was clearly separated at float precision because the
  float-truncated seed perturbed the solve enough to see rounds effect).
  Tolerance changed from strict e3 < e1 to e3 < e1 + 1e-6 m.

Closes #351

+++ libs/reusex/include/segmentation/Surfel.hpp
@@ -18,12 +18,13 @@ namespace reusex::geometry {
 /// The registration optimizer refines `world_pose` while the points/normals
 /// stay fixed in the optical frame. `world_pose` is the composition
 /// worldTf * localTf, i.e. the seed equals to_affine(pose) *
-/// to_affine(intrinsics.local_transform).
+/// to_affine(intrinsics.local_transform). Held in double to preserve the
+/// double-precision poses stored in ProjectDB end-to-end.
 struct FrameSurfels {
   int node_id = -1;
   CloudPtr points;   ///< pcl::PointXYZRGB in optical frame
   CloudNPtr normals; ///< pcl::Normal in optical frame (unit, facing camera)
-  Eigen::Affine3f world_pose = Eigen::Affine3f::Identity();
+  Eigen::Affine3d world_pose = Eigen::Affine3d::Identity();
 };
 
 } // namespace reusex::geometry
